Update LICENSE to GPLv3
Update the license of the template to better reflect the actual
licensing situation.

Due to how the GPL handles linking and the resulting binaries, all
binary (i.e. built) Jellyfin plugins will be GPLv3 licensed, as this is
the license of the binary NuGet packages that the plugin links against
at build time.

To avoid any potential confusion for plugin authors, update this license
so that the code license matches the license of the final binaries.

If a plugin author chooses to use another license, for instance a
permissive license so that the plugin can be linked/built against both
Emby and Jellyfin (insofar as that is possible technically), that is
fine and they should replace this LICENSE file in their plugin with one
reflecting the actual license chosen.